requirements.txt
psycopg2-binary>=2.9
pandas>=2.0
scikit-learn>=1.3
mcp>=1.0
